+++ flink-table/flink-table-planner/src/test/resources/org/apache/flink/table/planner/plan/batch/sql/join/hints/ShuffleMergeJoinHintTest.xml
